+++ lib/Phile/ServiceLocator/MetaInterface.php
@@ -20,5 +20,13 @@ interface MetaInterface
      *
      * @return array with key/value store
      */
-    public function parse($rawData);
+    public function extractMeta($rawData);
+
+    /**
+     * Parses text and extracts the content-part (meta-data is removed)
+     *
+     * @param string $rawData Text to inspect
+     * @return string Text without meta-data
+     */
+    public function extractContent(?string $rawData): string;
 }