+++ plugins/phile/parserMeta/tests/MetaTest.php
@@ -24,8 +24,22 @@ EOF;
             'fences' => ['yaml' => ['open' => '---', 'close' => '---']],
             'format' => 'YAML'
         ]);
-        $meta = $parser->parse($raw);
+        $meta = $parser->extractMeta($raw);
         $this->assertSame('foo', $meta['title']);
         $this->assertSame(['bar', 'baz'], $meta['tags']);
+    }
+
+    public function testDataWithoutMetaDataBlock()
+    {
+        $raw = "# Hello World\n## Hello you too";
+        $parser = new Meta([
+            'fences' => ['yaml' => ['open' => '---', 'close' => '---']],
+        ]);
+
+        $content = $parser->extractContent($raw);
+        $this->assertSame($raw, $content);
+
+        $meta = $parser->extractMeta($raw);
+        $this->assertSame([], $meta);
     }
 }